Change to get_value and set_value
You can get away with just saying PyAttributeA = PyAttributeB a lot
of times... but I think is more appropriate to return the string from
the get_value and then send the string to the set_value.

Not all PyAttributes are strings though... some are PyTime objects.  But
not in this case using PySegments.