Allow for PyTransitions in selection
If the artist does a box or shift selection from left to right on a
timeline, the selection is going to include the transition points.

Rather than using Select and Select in the Timeline to be certain that
only segments are selected, accomodate for PyTransitions.  PyTransitions
have no name metadata, so no harm to skip these.